changes in structure of project
