+++ app/dataset creation scripts/real.py
@@ -1,10 +1,6 @@
-import cv2
-import sys
+import cv2,os
 import random
-import os
-cascPath = "haarcascade_frontalface_default.xml"
-faceCascade = cv2.CascadeClassifier(cascPath)
-
+faceCascade = cv2.CascadeClassifier("haarcascade_frontalface_default.xml")
 video_capture = cv2.VideoCapture(0)
 
 while True:
@@ -17,6 +13,7 @@ while True:
         minNeighbors=5,
         minSize=(30, 30),
     )
+    
     cv2.rectangle(frame, (400, 100), (900, 550), (255,0,0), 2)
     # Draw a rectangle around the faces
     for (x, y, w, h) in faces:
@@ -24,15 +21,19 @@ while True:
         cv2.rectangle(frame, (x, y), (x+w, y+h), (0, 255, 0), 2)
         print(x,y,x+w,y+h)
         if(x<800 and x>400 and y<300 and y>100 and (x+w)<900 and (x+w)>400 and (y+h)<560 and (y+h) and len(faces)==1):
-            cv2.putText(frame,"Perfect", (10, 30),cv2.FONT_HERSHEY_SIMPLEX, 0.7, (0, 0, 255), 2)
             #if the face coems inside and teh dimensionsof the video is more than half the size of window
             # the frame will be captured.
-            cv2.imwrite(os.path.join('{path to directory of fake images}',"fake%d.jpg" % random.randint(40000,90000)), var1) 
+            cv2.putText(frame,"Perfect", (10, 30),cv2.FONT_HERSHEY_SIMPLEX, 0.7, (0, 0, 255), 2)
+            cv2.imwrite(os.path.join('{path to directory of real images}',"real%d.jpg" % random.randint(40000,90000)), var1) 
+
+  
     label = "{}".format(len(faces))
+    #cv2.putText(frame,label, (10, 30),cv2.FONT_HERSHEY_SIMPLEX, 0.7, (0, 0, 255), 2)
     cv2.imshow('Video', frame)
     # to show frames consecutively
     if cv2.waitKey(1) & 0xFF == ord('q'):
-        break     #on pressing the 'q' button the frame capturing will end.
+        break#on pressing the 'q' button the frame capturing will end.
 
+# When everything is done, release the capture
 video_capture.release()
 cv2.destroyAllWindows()
