+++ app/main.py
@@ -1,0 +1,8 @@
+from fastapi import FastAPI
+
+app=FastAPI()
+
+
+if __name__=="main":
+    predictperson()
+
